pull out functions from notebook
